Name attribute added to User SignUp wire Up and shows in AWS
